remove warning

+++ r2bix/src/p2048mini/p2048mini_Stage.h
@@ -33,7 +33,7 @@ namespace p2048mini
 		uint32_t Get( const uint32_t linear_index ) const;
 		uint32_t Get( const uint32_t x, const uint32_t y ) const;
 		int32_t GetCurrentNumberCount() const { return mCurrentNumberCount; }
-		int32_t GetEmptySpaceCount() const { return mContainer.size() - GetCurrentNumberCount(); }
+		int32_t GetEmptySpaceCount() const { return static_cast<int32_t>( mContainer.size() ) - GetCurrentNumberCount(); }
 		bool IsFull() const { return mContainer.size() == mCurrentNumberCount; }
 		bool IsMovable() const;
